feat: create Ui
